Revert "Start create fetch_and_validate method"
This reverts commit 3002746faacce55d9885c29447786f4f73c6cc4e.
